--- src/components/Dashboard/PostViewer/index.style.ts
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-export const Container = styled.div`
-  width: 94.3%;
-  height: auto;
-  margin: auto;
-`;
