+++ src/components/Post/List/index.style.ts
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const Container = styled.div`
+  width: 36.495%;
+  height: auto;
+  margin: auto;
+  padding-top: 50px;
+  padding-bottom: 100px;
+`;
