+++ src/components/Viewer/index.style.ts
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const Container = styled.div`
+  padding: 0 20px;
+  height: auto;
+  margin: auto;
+`;
